requirements.txt
django              == 2.2.0
psycopg2-binary     == 2.8.4
djangorestframework == 3.10.3
redis               == 3.4.1
pytest              == 3.6.3
pytest-django       == 3.3.3
django-redis        == 4.12.1

